fix: price 타입을 Long으로 변경 및 기본 생성자, getter 구현

+++ src/main/java/gift/domain/Product.java
@@ -4,10 +4,10 @@ public class Product {
 
     private Long id;
     private String name;
-    private Double price;
+    private Long price;
     private String imageUrl;
 
-    public Product(Long id, String name, Double price, String imageUrl) {
+    public Product(Long id, String name, Long price, String imageUrl) {
         this.id = id;
         this.name = name;
         this.price = price;
@@ -22,7 +22,7 @@ public class Product {
         return name;
     }
 
-    public Double getPrice() {
+    public Long getPrice() {
         return price;
     }
